add partial tests for grid filter
