+++ packages/ramp-core/package.json
@@ -74,6 +74,7 @@
         "autoprefixer": "^9.8.6",
         "cross-env": "^7.0.2",
         "cypress": "^6.4.0",
+        "cypress-plugin-tab": "~1.0.5",
         "eslint-plugin-prettier": "^3.1.1",
         "eslint-plugin-vue": "^6.0.0",
         "eslint": "^5.16.0",
